requirements.txt
cycler==0.10.0
decorator==4.4.2
easyocr==1.3.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
mysql-connector-python==8.0.25
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
protobuf==3.17.3
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tifffile==2021.6.6
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
